+++ examples/helper/mod.rs
@@ -1,9 +1,11 @@
-
 use bevy::prelude::*;
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
 
 mod camera;
 mod map;
+
+#[cfg(feature = "rapier")]
+pub mod rapier;
 
 #[derive(Default)]
 pub struct HelperPlugin;
@@ -13,5 +15,7 @@ impl Plugin for HelperPlugin {
         app.add_plugins(WorldInspectorPlugin::new());
         app.add_systems(Update, camera::movement);
         app.add_systems(Update, map::rotate);
+        #[cfg(feature = "rapier")]
+        app.add_systems(Update, rapier::move_player);
     }
 }